+++ Runtime/Core/Http/IHttpClient.cs
@@ -73,6 +73,7 @@ namespace AccelByte.Core {
     public interface IHttpRequestSender
     {
         void AddTask(IHttpRequest request, Action<HttpSendResult> callback, int timeoutMs, uint delayTimeMs);
+        void ClearTasks();
         void ClearCookies(Uri baseUri);
     }